Transitive Chains (Differ)
